Fix facebook login button behavior
When the user went to the signup page and came back to login, the button
would remain disabled, for it depended on the call of `fbAsyncInit`.
This function is only called when the facebook plugin script is added to
the page. Since it was already there, the call would never happen.